вспомогательные функции вынесены в отдельные модули
